Failure: g-i-installation_debian_sid_daily_hurd_lxde
Hello,

hurd g-i keeps failing just because some snapshots happen to be the
same just because the screen is (as expected) sometimes completely
blue.  As I mentioned in a previous mail, jenkins should be checking
all snapshots, not just one in the past.  I have attached a patch to
this mail which should be doing this.  I unfortunately don't have the
infrastructure to test it though, so there may be dragons.

Samuel