new rebootstrap job variants
The _supported variant uses the cross build method preferred by the gcc
maintainer. It relies on dpkg-cross instead of cross-gcc-dev. Adding
this variant, as it is not clear which one is to be used in the long
run. See: #766708

The _debbindiff variant compares the cross-built packages against their
in-archive natively-built counterparts. This processing takes
significantly longer, but can highlight errors in cross compilation.

This commit triples the number of rebootstrap jobs.