reproducible: Don't open external resources in our iframe
It's considered poor web etiquette to hijack links to cross-domain
resources - the default "click" should take me to, eg. bugs.debian.org, and
it can be quite annoying when it doesn't and I need to renavigate to get
the URL, etc.

Futhermore, when the iframe goes away we will have this behaviour anyway,
so not only can we do the right thing now, we can avoid changing habits in
the future after people are used or working around it.

Signed-off-by: Chris Lamb <lamby@debian.org>