reproducible debian: show the "show all" button when summary tag supported
The "summary" tag is presentally not supported in firefox... but will be with
the release of firefox version 48 in August 2016! In the meantime, this
commit hides the "see all package sets in suite/arch" button on the package
set pages used to expand all the summary tags for firefox.

Once Firefox 48 is released, this commit should be reverted.

Signed-off-by: Holger Levsen <holger@layer-acht.org>