reproducible debian: handle new packages with multiple versions
This commit fixes a bug introduced by 07849060

Commit 07849060 changed the UPSERT that added new packages to the sources
table into an INSERT. If the new package has two versions for (suite, arch)
in debian archive, then the INSERT to sources will fail on unique constraint.
The changes in this commit provide the same functionality without the UPSERT:
there will be no failure, instead only the second source encountered in
archive will be saved.

Signed-off-by: Holger Levsen <holger@layer-acht.org>