fix debootstrap fallout
jessie's debootstrap would chmod the target directory to 755, but the
bpo debootstrap no longer does that, so we need to chmod ourselves or
the / inside the chroot will be 700 (from mkdtemp).

The bpo debootstrap defaults to --merged-usr, but --merged-usr is broken
(see e.g. #810499). Revert until that is sorted out.

Signed-off-by: Holger Levsen <holger@layer-acht.org>