lvc: generally discard snapshots
Keeping them results in subsequent runs doing nothing much, which
is a little confusing.  We should instead separate jobs that start
from the snapshots, and then pass the snapshots between jobs when
they succeed, as this should make later tests more resilient to when
the early part is failing and would prevent the later tests from running.

Signed-off-by: Holger Levsen <holger@layer-acht.org>