reproducible FreeBSD: set pkg base timestamp
Packaged base is the work-in-progress distribution method for
FreeBSD 12.x, and is not yet completely connected into the FreeBSD
release build mechanism. For now explicitly set the pkg metadata
timestamp via SOURCE_DATE_EPOCH in the environment, until we have the
mechanism to set it via the release build scripts.

Signed-off-by: Holger Levsen <holger@layer-acht.org>