reproducible_lede: create a workaround for signing keys
LEDE signs the release with a signing key, but generate the signing key if not
present. To have a reproducible release we need to take care of signing keys.

LEDE will also put the key-build.pub into the resulting image (pkg: base-files)!
At the end of the build it will use the key-build to sign the Packages repo list.
Use a workaround this problem:

key-build.pub contains the pubkey of LEDE buildbot
key-build     contains our build key

Meaning only signed files will be different but not the images.
Packages.sig is unreproducible.

Signed-off-by: Holger Levsen <holger@layer-acht.org>