Move some IRC announcements to #debian-reproducible-changes.
This changes the behaviour such that we always announce on our "noisy"
#debian-reproducible-changes but keeps the main #debian-reproducible
channel free for a) when we really want to notify the channel such as when
we want to preserve artifacts or explicitly ask to be notified and b)
regular chat.

Otherwise, regular/vanilla reschedling can have the effect of masking other
conversations.

Signed-off-by: Chris Lamb <lamby@debian.org>
Signed-off-by: Holger Levsen <holger@layer-acht.org>