bin/find_dpkg_trigger_cycles.sh: update for Debian Stretch
 - Update from apt-file 2 to apt-file 3
     - the --cache and --sources-list are replaced by --config-file
     - "apt-file update" is replaced by "apt update"
     - apt config gains Acquire::IndexTargets section
 - Make use of features of more recent apt version
     - use sources.list in deb822 format
     - instead of relying on an uncompressed Packages file with a fixed
       pattern in /var/lib/apt/lists/, use "apt-get indextargets" in
       connection with "apt-helper cat-file"

Interface change:

Before, the script allowed either a suite name or codename to specify
the Debian release to check. Now, the script is limited to the codename
(jessie, stretch, buster, ...). The reason is, that we want to make sure
to grab the right Packages file using "apt-get indextargets" and for
that we need to specify the suite or the codename. To avoid having to
implement a mapping mechanism, the script now is limited to working with
codenames.

Signed-off-by: Holger Levsen <holger@layer-acht.org>